refactor Box components used to display Lift Stats and Targets
